Fix models mapping

+++ src/main/java/cz/charwot/chasify/models/User.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(nullable = false, unique = true)
     private String username;
@@ -48,10 +48,10 @@ public class User {
     @OneToMany(mappedBy = "user")
     private List<Activity> activities;
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
